Clean old file and new example
